requirements/common.txt
Click==7.0
coverage==4.5.1
decorator==4.4.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Script==2.0.6
Flask-WTF==0.14.2
infinity==1.4
intervals==0.8.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pillow==6.0.0
python-dotenv==0.10.2
six==1.12.0
validators==0.12.6
visitor==0.1.3
Werkzeug==0.15.4
WTForms==2.2.1
WTForms-Components==0.10.4
